Making importbox.sh and Vagrantfile as generic ones (applicable to every CIP testing project)
Trying to adopt the generic strategy to ANY Virtual Machine from vagrant repository, in order to make testing generic.

It is an attempt. The whole environment as such should act genericly on both VM providers: libvirt (integrated into kernel) and virtualbox (as independent ORACLE provider).

+++ importbox.sh
@@ -20,36 +20,73 @@ RED='\e[1;31m'
 NC='\e[0m' # No Color
 
 ## Start the vagrant assignment algorithm
-if [ $# -ne 2 ] ; then
-	echo "Usage: $0 boxNewName pathToBoxLocation"
-	echo "Example: ./importbox.sh debian_stretch64 debian/stretch64"
-	exit 1
+if [ $# -ne 1 ] && [ $# -ne 2 ] ; then
+    echo "Usage: $0 boxName (boxName => genericName)"
+    echo "Usage: $0 boxName pathToBoxLocation"
+    exit 1
 fi
 ## set -e
-boxNewName=$1
-pathToBoxLocation=$2
-echo -e "${RED}[1] Executing the command: vagrant box remove $boxNewName${NC}"
-vagrant box remove $boxNewName
+boxName=$1
+pathToBoxLocation=""
+if [ $# -eq 2 ] ; then
+    pathToBoxLocation=$2
+fi
+echo -e "${RED}[1] Executing the command: vagrant box remove $boxName${NC}"
+vagrant box remove $boxName
+set -e
 echo -e "${RED}[2] Executing the command: vagrant box list [to check which VMs remain]${NC}"
 vagrant box list
-echo -e "${RED}[3] Executing the command: vagrant box add $boxNewName $pathToBoxLocation${NC}"
-vagrant box add $boxNewName $pathToBoxLocation
+echo -e "${RED}[3] Executing the command: vagrant box add $boxName $pathToBoxLocation${NC}"
+vagrant box add $boxName $pathToBoxLocation
 if [ -f Vagrantfile ]; then
-    rm Vagrantfile
+    rm Vagrantfile ## rm old Vagrantfile, if any?
     ls -al
 fi
 
-echo -e "${RED}[4] Executing the command: vagrant init $boxNewName${NC}"
-vagrant init $boxNewName
-cp Vagrantfile.genesis Vagrantfile
-## sed -i Vagrantfile -e 's/debian\/stretch64/'"$box"'/'
+echo -e "${RED}[4] Executing the command: vagrant init $boxName${NC}"
+vagrant init $boxName
+ls -al Vagrantfile
+
+read -p "Do you want to overrwrite Vagrantfile with already provided generic Vagrantfile? [Y/n]" qm
+
+if [ -Z $qm ] || [ Y -eq $qm ] || [y -eq $qm ] ; then
+    echo "Overwrite Vagrant file"
+    cp Vagrantfile.genesis Vagrantfile
+    sed -i Vagrantfile -e 's/boxName/'"$boxName"'/'
+    ls -al Vagrantfile
+else
+    echo "Keep original Vagrantfile"
+    ls -al Vagrantfile
+fi
+
 echo -e "${RED}[5] Executing the command: vagrant box list [to check if new VM is added]${NC}"
 vagrant box list
 
 # are any mods necessary?
-## set +e
+set +e
 echo -e "${RED}[6] Executing the command: vagrant up --provider virtualbox${NC}"
-vagrant up --provider virtualbox
+echo "==> Available vm providers:"
+echo "1) libvirt"
+echo "2) virtualbox"
+read -p "==> Which vm provider should be used? " provider
+
+case $provider in
+[1])
+    echo "The vm provider is libvirt"
+    myprovider="libvirt"
+    ;;
+
+[2])
+    echo "The vm provider is virtualbox"
+    myprovider="virtualbox"
+    ;;
+
+*)
+    echo "Did not give the correct vm provider!"
+    ;;
+esac
+
+vagrant up --provider $myprovider
 echo The above should end with "==> default: KernelCI already configured remove ~/mybbb.dat to force configuration"
 echo and a report that the ssh command responded with a non-zero exit status. This is expected!
 echo
